fix: replace sql query for films

+++ src/main/resources/schema.sql
@@ -7,12 +7,12 @@ DROP TABLE if exists users cascade;
 DROP TABLE if exists films cascade;
 
 CREATE TABLE IF NOT EXISTS mpa (
-                                   id integer not null PRIMARY KEY auto_increment,
-                                   name varchar(255)
+                       id bigint not null PRIMARY KEY auto_increment,
+                       name varchar(255)
 );
 
 CREATE TABLE IF NOT EXISTS films (
-                        id integer PRIMARY KEY auto_increment,
+                        id bigint PRIMARY KEY auto_increment,
                         name varchar(255),
                         description varchar(200),
                         releaseDate date,
@@ -21,7 +21,7 @@ CREATE TABLE IF NOT EXISTS films (
 );
 
 CREATE TABLE IF NOT EXISTS users (
-                        id integer not null PRIMARY KEY auto_increment,
+                        id bigint not null PRIMARY KEY auto_increment,
                         email varchar(255) not null UNIQUE,
                         login varchar(255) not null UNIQUE,
                         name varchar(255),
@@ -29,25 +29,25 @@ CREATE TABLE IF NOT EXISTS users (
 );
 
 CREATE TABLE IF NOT EXISTS genres (
-                        id integer not null PRIMARY KEY auto_increment,
+                        id bigint not null PRIMARY KEY auto_increment,
                         name varchar(255)
 );
 
 CREATE TABLE IF NOT EXISTS film_genres (
-                        film_id integer references films(id),
-                        genre_id integer references genres(id),
+                        film_id bigint references films(id),
+                        genre_id bigint references genres(id),
                         primary key (film_id, genre_id)
 );
 
 CREATE TABLE IF NOT EXISTS likes (
-                         user_id integer references users(id),
-                         film_id integer references films(id),
+                         user_id bigint references users(id),
+                         film_id bigint references films(id),
                          primary key (user_id, film_id)
 
 );
 
 CREATE TABLE IF NOT EXISTS friends (
-                           user_id integer references users(id),
-                           friend_id integer references users(id),
+                           user_id bigint references users(id),
+                           friend_id bigint references users(id),
                            primary key (user_id, friend_id)
 );
